Move checker functions to more suited place
